+++ src/ft_vector_draw.c
@@ -6,22 +6,22 @@
 /*   By: msamhaou <msamhaou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/21 17:24:33 by msamhaou          #+#    #+#             */
-/*   Updated: 2023/06/25 03:34:00 by msamhaou         ###   ########.fr       */
+/*   Updated: 2023/06/25 04:13:13 by msamhaou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-double	*ft_hypo_calc(t_vertex *a, t_vertex *b)
+float	*ft_hypo_calc(t_vertex *a, t_vertex *b)
 {
-	double	*hypo;
-	double	x;
-	double	y;
-	double	deg;
+	float	*hypo;
+	float	x;
+	float	y;
+	float	deg;
 
 	x = (b->x - a->x);
 	y = (a->y - b->y);
-	hypo = malloc(sizeof(double) * 2);
+	hypo = malloc(sizeof(float) * 2);
 	hypo[0] = sqrt(pow(x, 2) + pow(y, 2));
 	hypo[1] = asin(y / hypo[0]);
 	return (hypo);
@@ -38,12 +38,12 @@ void	ft_swap_vect(t_vertex **a, t_vertex **b)
 
 void	ft_vect_draw(t_vertex *a, t_vertex *b, int color, t_data *data)
 {
-	double	i;
-	double	*hypo;
-	double	deg;
-	double	adj;
-	double	x;
-	double	y;
+	float	i;
+	float	*hypo;
+	float	deg;
+	float	adj;
+	float	x;
+	float	y;
 
 	i = 0;
 	hypo = ft_hypo_calc(a, b);
@@ -57,7 +57,7 @@ void	ft_vect_draw(t_vertex *a, t_vertex *b, int color, t_data *data)
 		else
 			x+= cos(hypo[1]);
 		y += sin(hypo[1]);
-		
+
 		i++;
 	}
 }
